+++ src/mantle/testing/concerns/trait-reads-annotations.php
@@ -92,4 +92,13 @@ trait Reads_Annotations {
 			...$method->getAttributes( $name ),
 		];
 	}
+
+	/**
+	 * Check if the method has an attribute of a given name.
+	 *
+	 * @param string $name The name of the method to check.
+	 */
+	public function method_has_attribute( string $name ): bool {
+		return ! empty( $this->get_attributes_for_method( $name ) );
+	}
 }